+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 built-in recipe generator). This ensures that changes to external files and others are correctly caught and handled.
 - Broke `Status.Dirty` into several more explicit causes of dirtiness (e.g. `OutputsInvalid`) to make it more clear to
 the user why a given recipe has been marked dirty
+- Moved functions for checking Recipe and ForeachRecipe dirtiness to alkymi.alkymi to reduce the complexity of the
+  classes
 
 ### Fixed
 - Converted several captured variables inside tests to globals to avoid them interfering with hashing of the bound